move APIHandler to Handler package
